% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dubois_key_gylph.R
\name{draw_key_dubois}
\alias{draw_key_dubois}
\title{Change legend symbols to filled circles}
\usage{
draw_key_dubois(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to display in this key.}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Scaled size of key circle}
}
\value{
A grid grob.
}
\description{
Change legend symbols to filled circles
}
\examples{
library(ggplot2)
p <- ggplot(economics, aes(date, psavert, fill = "savings rate"))
# key glyphs can be specified by their name
p +
  geom_point(
    shape = 21,
    key_glyph = "dubois")

# key glyphs can be specified via their drawing function
p +
  geom_point(
    shape = 21,
    key_glyph = draw_key_dubois)
}
