% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dubois_colors.R
\name{dubois_color}
\alias{dubois_color}
\title{Colors used by Dubois}
\usage{
dubois_color(...)
}
\arguments{
\item{...}{Color choice or choices}
}
\value{
Vector of colors
}
\description{
Colors used by Dubois
}
\examples{
dubois_color("crimson")
dubois_color("brown", "blue", "pink")
}
