% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dubois_colors.R
\name{dubois_palette}
\alias{dubois_palette}
\title{Palettes of Dubois's colors}
\usage{
dubois_palette(palette = "main")
}
\arguments{
\item{palette}{Choice of palette: "all", "main", "alt", "seq", "seq2", "seq3", "seq4", "div", "div2", "div3", "highlight", "highlight2", highlight3"}
}
\value{
list of colors
}
\description{
Palettes of Dubois's colors
}
\examples{
dubois_palette()
dubois_palette("alt")
}
