% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_grid.R
\name{ggplot_add.conditional_grid}
\alias{ggplot_add.conditional_grid}
\title{crafts conditional grid}
\usage{
\method{ggplot_add}{conditional_grid}(object, plot, object_name)
}
\description{
crafts conditional grid
}
