% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dubois_colors.R
\name{scale_fill_dubois}
\alias{scale_fill_dubois}
\title{Fill palette of Dubois colors}
\usage{
scale_fill_dubois(palette = "main", direction = 1, order = NULL, ...)
}
\arguments{
\item{palette}{Choice of palette name or number}

\item{direction}{Direction of color choices}

\item{order}{Order of color choices}

\item{...}{Additional parameters passed internally}
}
\value{
ggplot object with colors applied
}
\description{
Fill palette of Dubois colors
}
\examples{
 library(dplyr)
 library(ggplot2)
 mtcars |>
   mutate(cyl = factor(cyl)) |>
   ggplot(aes(mpg, cyl, fill = cyl)) +
   geom_boxplot(key_glyph = "dubois") +
   theme_mini(conditional = TRUE) +
   scale_fill_dubois()
}
