% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dubois_theme.R
\name{theme_dubois}
\alias{theme_dubois}
\title{Minimal Dubois-style theme, with easy customization}
\usage{
theme_dubois(bg = dubois_color("linen"), minor = FALSE, conditional = TRUE)
}
\arguments{
\item{bg}{Background color}

\item{minor}{Minor gridlines}

\item{conditional}{Conditional gridlines}
}
\value{
ggplot object
}
\description{
Minimal Dubois-style theme, with easy customization
}
\examples{
 library(dplyr)
 library(ggplot2)
 mtcars |>
   mutate(cyl = factor(cyl)) |>
   ggplot(aes(mpg, cyl, fill = cyl)) +
   geom_boxplot(key_glyph = "dubois") +
   theme_dubois() +
   scale_fill_dubois()
}
