% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mini.R
\name{theme_mini}
\alias{theme_mini}
\title{Simple, minimal theme, with easy customization}
\usage{
theme_mini(bg = "white", text = "black", minor = FALSE, conditional = TRUE)
}
\arguments{
\item{bg}{Background color}

\item{text}{Not used}

\item{minor}{Minor gridlines}

\item{conditional}{Conditional gridlines}
}
\value{
ggplot object
}
\description{
Simple, minimal theme, with easy customization
}
\examples{
 library(dplyr)
 library(ggplot2)
 mtcars |>
   mutate(cyl = factor(cyl)) |>
   ggplot(aes(mpg, cyl, fill = cyl)) +
   geom_boxplot(key_glyph = "dubois") +
   theme_mini() +
   scale_fill_dubois()

 mtcars |>
   mutate(cyl = factor(cyl)) |>
   ggplot(aes(cyl, mpg, fill = cyl)) +
   geom_point(shape = 21, size = 4) +
   theme_mini(bg = "#ddddff")
}
