% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viridis_palettes.R
\name{viridis_palettes}
\alias{viridis_palettes}
\title{Display Viridis palettes}
\usage{
viridis_palettes(palettes = 1:8, colors = 9)
}
\arguments{
\item{palettes}{Choice of palettes, corresponding to one of the 8 palettes included with Viridis}

\item{colors}{The number of colors to generate for each palette}
}
\value{
ggplot2 object
}
\description{
Display Viridis palettes
}
\examples{
viridis_palettes()
}
