% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mini.R
\name{theme_mini}
\alias{theme_mini}
\title{Simple, minimal theme, with easy customizations}
\usage{
theme_mini(bg = "white", text = "black", minor = FALSE, conditional = TRUE)
}
\arguments{
\item{bg}{Background color}

\item{text}{Not used}

\item{minor}{Minor gridlines}

\item{conditional}{Conditional gridlines}
}
\value{
ggplot object
}
\description{
Simple, minimal theme, with easy customizations
}
\examples{
 library(dplyr)
 library(ggplot2)
 mtcars |>
   mutate(cyl = factor(cyl)) |>
   ggplot(aes(mpg, cyl, fill = cyl)) +
   geom_boxplot(key_glyph = "dubois") +
   theme_mini(conditional = TRUE) +
   scale_fill_dubois()
}
